cryage_backend/requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.32.0
python-socketio>=5.11.4
redis>=5.2.0
httpx>=0.27.2
pandas>=2.2.3
pydantic>=2.9.2
python-dotenv>=1.0.1
agno>=1.4.5
pydantic-settings
